requirements.txt
# ImaraFund Backend Requirements

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25

# Data Validation  
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# AI Integration (Gemini 2.5 Flash)
google-generativeai==0.3.2

# Database Drivers
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.3

# API Utilities
python-multipart==0.0.6
